Retrieve GroupNode's underlying input- or output-splitters
